adding --gke flag, some related cleanup and restructuring
